# spec/openapi.yaml
openapi: "3.0.3"
info:
  title: MOC Onboarding Microservice API
  description: >-
    Common API to link oboarding system to services such as
    OpenStack and OpenShift.
  contact:
    url: https://github.com/larsks/acct-manager
  version: "20211215"
servers:
  - url: http://localhost:8080
    description: "Development server"
  - url: https://onboarding-onboarding.apps-crc.testing
    description: "CRC Deployment"

##
## PATHS
##

paths:
  /healthz:
    description: >-
      Health check endpoint; this endpoint does not require authentication.
    get:
      operationId: healthcheck
      tags:
        - healthcheck
      summary: Check service health
      responses:
        "200":
          description: Service is running
          content:
            text/plain:
              schema:
                type: string
                example: OK
  /projects:
    post:
      operationId: createProject
      tags:
        - project
      summary: Create a new project
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "definitions.yaml#/definitions/ProjectRequest"
      responses:
        "200":
          description: Project creation was successful
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/ProjectResponse"
        "409":
          description: Conflict with existing project name
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
  /projects/{project_name}:
    parameters:
      - name: project_name
        in: path
        schema:
          type: string
        required: true
    get:
      operationId: getProject
      tags:
        - project
      summary: Get information about a project
      security:
        - basicAuth: []
      responses:
        "200":
          description: Project exists
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/ProjectResponse"
        "404":
          description: Project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    delete:
      operationId: deleteProject
      tags:
        - project
      summary: Delete a project
      security:
        - basicAuth: []
      responses:
        "200":
          description: Project was deleted successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "403":
          description: Attempt to operate on forbidden project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: Project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
  /projects/{project_name}/quotas:
    parameters:
      - name: project_name
        in: path
        schema:
          type: string
        required: true
    get:
      operationId: getProjectQuota
      tags:
        - quota
      summary: Get information about project quotas
      security:
        - basicAuth: []
      responses:
        "200":
          description: Project quota information
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/QuotaResponse"
        "403":
          description: Attempt to access invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: Project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    put:
      operationId: updateProjectQuota
      tags:
        - quota
      summary: Update a project quota
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "definitions.yaml#/definitions/QuotaRequest"
      responses:
        "200":
          description: Project quota updated successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/QuotaResponse"
        "403":
          description: Attempt to access invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: Project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    delete:
      operationId: deleteProjectQuota
      tags:
        - quota
      summary: Delete a project quota
      security:
        - basicAuth: []
      responses:
        "200":
          description: Project quota deleted successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "403":
          description: Attempt to access invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: Project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
  /users:
    post:
      operationId: createUser
      tags:
        - user
      summary: Create a user
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "definitions.yaml#/definitions/UserRequest"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/UserResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
  /users/{user_name}:
    parameters:
      - name: user_name
        in: path
        schema:
          type: string
        required: true
    get:
      operationId: getUser
      tags:
        - user
      summary: Get information about a user
      security:
        - basicAuth: []
      responses:
        "200":
          description: User exists
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/UserResponse"
        "404":
          description: User does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    delete:
      operationId: deleteUser
      tags:
        - user
      summary: Delete a user
      security:
        - basicAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: User does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
  /users/{user_name}/projects/{project_name}/roles/{role_name}:
    parameters:
      - name: user_name
        in: path
        schema:
          type: string
        required: true
      - name: project_name
        in: path
        schema:
          type: string
        required: true
      - name: role_name
        in: path
        schema:
          type: string
        required: true
    get:
      operationId: getUserRole
      tags:
        - role
      summary: Check if user has been granted role
      security:
        - basicAuth: []
      responses:
        "200":
          description: Information about role membership
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/RoleResponse"
        "403":
          description: Attempt to access an invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: User or project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    put:
      operationId: grantUserRole
      tags:
        - role
      summary: Grant role to user on project
      security:
        - basicAuth: []
      responses:
        "200":
          description: Role granted successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "403":
          description: Attempt to access an invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: User or project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
    delete:
      operationId: revokeUserRole
      tags:
        - role
      summary: Revoke role for user on project
      security:
        - basicAuth: []
      responses:
        "200":
          description: Role revoked successfully
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "403":
          description: Attempt to access an invalid project
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "404":
          description: User or project does not exist
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"
        "400":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "definitions.yaml#/definitions/Response"

##
## COMPONENTS
##

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

##
## SECURITY
##

security:
  - basicAuth: []